requirements.txt
pandas==2.0.2
numpy==1.24.3
numba==0.59.1
pytest==8.0.0
python-binance==1.0.19
plotly==5.18.0
imageio==2.31.6
Pillow==10.0.0
pytz==2024.1 
python-telegram-bot==20.0
schedule>=1.1.0